worker/requirements.txt
google-genai
psycopg[binary]
python-dotenv
